removing auto layout functionality
over time this logic has gotten in the way more than it has helped.

e.g. witha request for "foo.json" the request handler is already trying
to render /views/layouts/json/default.ctp

by then setting the layout to a different file (becuase isAjax will
return true) the code ends up trying to render

 /views/layouts/json/ajax.ctp

It was written at a time when the choice was either ajax or html - that
isn't the case in the projects where this code is deployed, and
therefore this method has been removed. If you used this method and
"wanted" it - check the requst handler docs as it probably does what you
want already.

The other side of the functionality (changing to admin_default.ctp for
admin requests) is handled much better using the mi_panel plugin.