allow setting post actions for other controllers
one of the failings with the user of postActions - is taht a link to a
postAction-ed action in another controller cannot be specified. This fix
permits setting
	$this->set('postActions', array('foo' => array('delete'));

anywhere in your code and these settings will persist. If you're in the
foo conttroller - you're overriding what the swiss army component will
do. if you're not, this is in addition to what the swiss army component
will do